Compare geotrellis raster objects

Evaluate whether two gt_RasterLayer objects have the same extent, number of rows and columns, projection, and resolution (or a subset of these comparisons).

gt_compareRaster(x, y, extent = TRUE, rowcol = TRUE, crs = TRUE,
  res = TRUE, tolerance = 0.1, stopiffalse = TRUE, showwarning = FALSE)

Arguments

x

gt_RasterLayer.

y

gt_RasterLayer.

extent

logical should the objects' extent be compared? Defaults to TRUE.

rowcol

logical should the objects' dimensionality be compared? Defaults to TRUE.

res

logical should the objects' resolution be compared? Defaults to TRUE.

tolerance

numeric This sets the difference between values that is permissible for them to be considered equal. Defaults to 0.1.

stopiffalse

logical should an error be thrown if one or more conditions are not met?

showwarning

logical should a warning be thrown if one or more conditions are not met?

Value

logical are all tested parameters the same?

Details

This function is similar to compareRaster except that is is less rigorous.