Reproject geotrellis raster data

This function reprojects a geotrellis raster data gt_RasterLayer object to a new coordinate system. This function is similar to projectRaster.

gt_projectRaster(from, to, ...)

Arguments

from

gt_RasterLayer object.

...

additional arguments.

crs

crs new coordinate system to project data to.

method

character method used to compute values for the new gt_RasterLayer. Either 'ngb' (nearest neighbor), which is useful for categorical variables, or 'bilinear' (bilinear interpolation; the default value), which is appropriate for continuous variables.

Value

gt_RasterLayer object.

Details

Note that--unlike projectRaster--this function does not have a to argument. Therefore users cannot specify a template projected raster to reproject a project gt_RasterLayer and ensure that the returned raster has a specific extent and dimensionality.