Create geotrellis raster object

This function creates a geotrellis raster object (gt_RasterLayer).

gt_raster(x, ...)

Arguments

x

input object.

...

additional arguments passed to raster if x is not character or RasterLayer.

Value

gt_RasterLayer object.

Details

This function essentially loads x into geotrellis as as raster data using the following methods:

character objectx is treated as a file path for raster data loaded directly into the Scala interpreter.

RasterLayer objectx is saved to disk and loaded into the Scala interpreter.

otherwisean attempt is made to coerce x to a RasterLayer object and then it is saved to disk and loaded into the Scala interpreter.

See also

raster.

Examples

# create gt_raster object using a matrix
x <- gt_raster(matrix(1:4, ncol=2))
# create gt_raster object using a raster
x <- gt_raster(raster(matrix(1:4, ncol=2)))
# create gt_raster object using a file path
x <- gt_raster(system.file("external/test.grd", package="raster"))

#> Error: File 'test.grd' does not have extensiontif